refactor: Move WebClient DTOs
